+++ package.json
@@ -52,7 +52,7 @@
     "db:seed:prod": "node_modules/.bin/sequelize db:seed:all --options-path .production.sequelizerc",
     "db:seed:undo": "node_modules/.bin/sequelize db:seed:undo:all",
     "db:seed:undo:prod": "node_modules/.bin/sequelize db:seed:undo:all --options-path .production.sequelizerc",
-    "docker:deps": "docker-compose run --rm backend yarn install && docker-compose run --rm frontend yarn install",
+    "docker:deps": "docker-compose run --rm backend yarn install && docker-compose run --rm frontend yarn install && docker-compose run --rm testingonly yarn install",
     "docker:reset": "./bin/reset-all",
     "docker:start": "docker-compose up",
     "docker:start:debug": "docker-compose --compatibility -f docker-compose.yml -f docker-compose.debug.yml up",
